+++ Dangerfile
@@ -20,6 +20,6 @@ checkstyle_format.report "app/build/reports/ktlint/ktlintMainSourceSetCheck.xml"
 # AndroidLint
 android_lint.report_file = "app/build/reports/lint-results.xml"
 android_lint.gradle_task = "runChecksForDanger"
-# android_lint.filtering = true
+android_lint.filtering = true
 # android_lint.severity = "Error"
 android_lint.lint(inline_mode: true)
