+++ app/build.gradle
@@ -10,7 +10,7 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.0"
     defaultConfig {
-        applicationId "topads.tokopedia.com.tinkersample"
+        applicationId "com.tokopedia.sample"
         minSdkVersion 16
         targetSdkVersion 29
         versionCode 1
