+++ app/src/main/java/com/tokopedia/sample/BaseApplicationLike.java
@@ -1,4 +1,4 @@
-package topads.tokopedia.com.tinkersample;
+package com.tokopedia.sample;
 
 import android.annotation.TargetApi;
 import android.app.Application;
