+++ app/src/main/java/com/tokopedia/sample/network/GetDataService.java
@@ -1,4 +1,4 @@
-package topads.tokopedia.com.tinkersample.network;
+package com.tokopedia.sample.network;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
