+++ app/src/main/java/com/tokopedia/sample/network/RetrofitClientInstance.java
@@ -1,4 +1,4 @@
-package topads.tokopedia.com.tinkersample.network;
+package com.tokopedia.sample.network;
 
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
