+++ app/src/main/java/com/tokopedia/sample/utils/ScreenUtils.kt
@@ -1,10 +1,8 @@
-package topads.tokopedia.com.tinkersample.utils
+package com.tokopedia.sample.utils
 
 import android.content.Context
 import android.util.DisplayMetrics
 import android.view.WindowManager
-
-
 /**
  * Author errysuprayogi on 09,January,2020
  */
@@ -25,7 +23,8 @@ object ScreenUtils {
 
     fun getStatusBarHeight(context: Context): Int {
         var result = 0
-        val resourceId: Int = context.getResources().getIdentifier("status_bar_height", "dimen", "android")
+        val resourceId: Int = context.getResources()
+            .getIdentifier("status_bar_height", "dimen", "android")
         if (resourceId > 0) {
             result = context.getResources().getDimensionPixelSize(resourceId)
         }
