+++ app/src/main/java/com/tokopedia/sample/utils/UserPrefs.kt
@@ -1,4 +1,4 @@
-package topads.tokopedia.com.tinkersample.utils
+package com.tokopedia.sample.utils
 
 import android.content.Context
 import android.content.SharedPreferences
@@ -9,7 +9,7 @@ import android.content.SharedPreferences
 class UserPrefs private constructor(private val context: Context) {
     private val settings: SharedPreferences
     fun saveString(key: String?, value: String?) {
-        settings.edit().putString(key, value).commit()
+        settings.edit().putString(key, value).apply()
     }
 
     fun getString(key: String?, defaultValue: String?): String? {
